Added ability to run only finishing script

+++ upgrade.sh
@@ -24,7 +24,7 @@ done
 
 if [ ! -z $help ]; then
 	echo "" | tee -a /tmp/upgradeLog
-	echo "Usage: $0 [--CONFIRM_ARCH] [--SKIP_CHECK] [--SKIP_CONFIRM]" | tee -a /tmp/upgradeLog
+	echo "Usage: $0 [--CONFIRM_ARCH] [--SKIP_CHECK] [--SKIP_CONFIRM] [--SKIP_INSTALL]" | tee -a /tmp/upgradeLog
 	echo "" | tee -a /tmp/upgradeLog
 	exit 0
 fi
@@ -113,7 +113,22 @@ fi
 # Exit if the latest version of Alpine Linux is already installed
 if [[ $VERSION_ID == $LATEST_RELEASE ]]; then
 	echo "Already the latest version of Alpine Linux. Move along, nothing to see here." | tee -a /tmp/upgradeLog
-	exit 0
+ 	while true; do
+		echo -n "Would you like to rerun the finishing script instead (y/n)? [n] " | tee -a /tmp/upgradeLog
+		read confirm <&1 # Read from stdout instead of stdin since it would just read next line of script when run through pipe
+		echo $confirm >> /tmp/upgradeLog
+		if [[ "$confirm" == "y" ]] || [[ "$confirm" == "Y" ]]; then
+			echo "User selected to proceed." | tee -a /tmp/upgradeLog
+   			SKIP_CHECK=
+      			SKIP_INSTALL=
+			break
+		elif [[ "$confirm" == "n" ]] || [[ "$confirm" == "N" ]] || [[ -z $confirm ]]; then
+			echo "User selected No." | tee -a /tmp/upgradeLog
+			exit 0
+		else
+			echo "Please answer yes (y) or no (n)." | tee -a /tmp/upgradeLog
+		fi
+	done
 fi
 
 if [ -z $SKIP_CHECK ]; then
@@ -194,6 +209,8 @@ fi
 
 #### Start Update
 
+if [ ! -z $SKIP_INSTALL ]; then
+
 # Comment out edge repositories so we are on latest stable build
 sed -e '/\/edge\// s/^#*/#/' -i /etc/apk/repositories 2>&1 | tee -a /tmp/upgradeLog
 
@@ -223,7 +240,11 @@ start_pre() {
                 /run/os-upgrade /var/log/os-upgrade
 }
 ' > /etc/init.d/os-upgrade
+
 chmod +x /etc/init.d/os-upgrade 2>&1 | tee -a /tmp/upgradeLog
+
+fi
+
 echo "#!/bin/ash
 # Alpine Linux OS Version Updater - Finishing Script
 # By: XtendedGreg - https://youtube.com/@XtendedGreg
@@ -237,10 +258,16 @@ echo 'Based on https://wiki.alpinelinux.org/wiki/Upgrading_Alpine' | tee -a /tmp
 echo '' | tee -a /tmp/upgradeLog
 echo \"Startup Date: \$(date)\" | tee -a /tmp/upgradeLog
 echo '' | tee -a /tmp/upgradeLog
+" > /bin/os-upgrade.sh
 
-echo 'Moved old repositories list to /etc/apk/repositories.bak' | tee -a /tmp/upgradeLog
-mv -v /etc/apk/repositories /etc/apk/repositories.bak 2>&1 | tee -a /tmp/upgradeLog
+if [ ! -z $SKIP_INSTALL ]; then
+	echo "
+	echo 'Moved old repositories list to /etc/apk/repositories.bak' | tee -a /tmp/upgradeLog
+	mv -v /etc/apk/repositories /etc/apk/repositories.bak 2>&1 | tee -a /tmp/upgradeLog
+	" >> /bin/os-upgrade.sh
+fi
 
+echo"
 # Copy over local repositories that are not commented out
 cat /etc/apk/repositories.bak | grep -v 'http' | grep -e -v '^#' > /etc/apk/repositories
 
@@ -260,9 +287,14 @@ apk cache clean 2>&1 | tee -a /tmp/upgradeLog
 # Correct packages that did not exist on upgrade
 apk fix 2>&1 | tee -a /tmp/upgradeLog
 
-rc-update del /etc/init.d/os-upgrade 2>&1 | tee -a /tmp/upgradeLog
-lbu exclude /etc/init.d/os-upgrade /bin/os-upgrade.sh 2>&1 | tee -a /tmp/upgradeLog
-rm /etc/init.d/os-upgrade /bin/os-upgrade.sh 2>&1 | tee -a /tmp/upgradeLog
+if [ -e /etc/init.d/os-upgrade ]; then
+	rc-update del /etc/init.d/os-upgrade 2>&1 | tee -a /tmp/upgradeLog
+ 	lbu exclude /etc/init.d/os-upgrade 2>&1 | tee -a /tmp/upgradeLog
+  	rm /etc/init.d/os-upgrade 2>&1 | tee -a /tmp/upgradeLog
+fi
+
+lbu exclude /bin/os-upgrade.sh 2>&1 | tee -a /tmp/upgradeLog
+rm /bin/os-upgrade.sh 2>&1 | tee -a /tmp/upgradeLog
 
 lbu commit 2>&1 | tee -a /tmp/upgradeLog
 
@@ -275,59 +307,66 @@ mount -o remount,rw /media/$LBU_MEDIA 2>&1 | tee -a /tmp/upgradeLog
 cat /tmp/upgradeLog >> /media/$LBU_MEDIA/upgradeLog
 rm /tmp/upgradeLog
 mount -o remount,ro /media/$LBU_MEDIA
-" > /bin/os-upgrade.sh
+" >> /bin/os-upgrade.sh
 chmod +x /bin/os-upgrade.sh 2>&1 | tee -a /tmp/upgradeLog
-lbu add /etc/init.d/os-upgrade /bin/os-upgrade.sh 2>&1 | tee -a /tmp/upgradeLog
 
-echo "Adding post-upgrade run-once script to RC" | tee -a /tmp/upgradeLog
-rc-update add os-upgrade default 2>&1 | tee -a /tmp/upgradeLog
+if [ ! -z $SKIP_INSTALL ]; then
 
-# Save final config before upgrade starts
-lbu commit 2>&1 | tee -a /tmp/upgradeLog
-
-cd /media/$LBU_MEDIA
-mount -o remount,rw /media/$LBU_MEDIA 2>&1 | tee -a /tmp/upgradeLog
-cat /tmp/upgradeLog >> upgradeLog
-rm /tmp/upgradeLog 2>&1 | tee -a upgradeLog
-wget https://dl-cdn.alpinelinux.org/alpine/latest-stable/releases/${ARCH}/alpine-rpi-${LATEST_RELEASE}-${ARCH}.tar.gz 2>&1 | tee -a upgradeLog
-wget https://dl-cdn.alpinelinux.org/alpine/latest-stable/releases/${ARCH}/alpine-rpi-${LATEST_RELEASE}-${ARCH}.tar.gz.sha256 2>&1 | tee -a upgradeLog
-if [ $(sha256sum -c alpine-rpi-${LATEST_RELEASE}-${ARCH}.tar.gz.sha256 | grep "alpine-rpi-${LATEST_RELEASE}-${ARCH}.tar.gz: OK" | wc -l) -eq 1 ]; then
-	echo "Alpine Linux Release checksum confirmed. Proceeding with upgrade..." | tee -a upgradeLog
- 	# Remove old files which will be replaced when new version is extracted
- 	rm -r /media/$LBU_MEDIA/apks 2>&1 | tee -a upgradeLog
-  	rm -r /media/$LBU_MEDIA/boot 2>&1 | tee -a upgradeLog
-   	rm -r /media/$LBU_MEDIA/overlays 2>&1 | tee -a upgradeLog
-    	rm /media/$LBU_MEDIA/*.dtb 2>&1 | tee -a upgradeLog
-     	rm /media/$LBU_MEDIA/*.elf 2>&1 | tee -a upgradeLog
-      	rm /media/$LBU_MEDIA/*.dat 2>&1 | tee -a upgradeLog
-       	rm /media/$LBU_MEDIA/bootcode.bin 2>&1 | tee -a upgradeLog
-	rm /media/$LBU_MEDIA/cmdline.txt 2>&1 | tee -a upgradeLog
- 	rm /media/$LBU_MEDIA/config.txt 2>&1 | tee -a upgradeLog
-	#rm /media/$LBU_MEDIA/apks/$ARCH/* # Clear old apk packages from previous version
-	#rm /media/$LBU_MEDIA/cache/*
-	#apk update | tee -a upgradeLog
-	#apk update | tee -a upgradeLog # Needs to be run twice?
-	#apk cache -v download | tee -a upgradeLog
-	tar xzf alpine-rpi-${LATEST_RELEASE}-${ARCH}.tar.gz 2>&1 | tee -a upgradeLog
-	rm alpine-rpi-${LATEST_RELEASE}-${ARCH}.tar.gz alpine-rpi-${LATEST_RELEASE}-${ARCH}.tar.gz.sha256 2>&1 | tee -a upgradeLog
-	echo "Upgrade Complete.  Syncing drive and rebooting..." | tee -a upgradeLog
-	echo "" | tee -a upgradeLog
-	sync 2>&1 | tee -a upgradeLog
-	reboot
-else
-	echo "DOWNLOADED FILE CHECKSUM FAILURE. ABORTING AND CLEANING UP." | tee -a upgradeLog
-	rm alpine-rpi-$LATEST_RELEASE-$ARCH.tar.gz alpine-rpi-$LATEST_RELEASE-$ARCH.tar.gz.sha256 2>&1 | tee -a upgradeLog
+	lbu add /etc/init.d/os-upgrade /bin/os-upgrade.sh 2>&1 | tee -a /tmp/upgradeLog
 	
-	rc-update del /etc/init.d/os-upgrade 2>&1 | tee -a upgradeLog
-	lbu exclude /etc/init.d/os-upgrade /bin/os-upgrade.sh 2>&1 | tee -a upgradeLog
-	rm /etc/init.d/os-upgrade /bin/os-upgrade.sh 2>&1 | tee -a upgradeLog
-
-	lbu commit 2>&1 | tee -a upgradeLog
-	echo "CLEANUP COMPLETE. EXITING WITH ERROR CODE 1." | tee -a upgradeLog
-	echo "View upgrade log: /media/$LBU_MEDIA/upgradeLog" | tee -a upgradeLog
-	echo "" | tee -a upgradeLog
-	mount -o remount,ro /media/$LBU_MEDIA
-	exit 1
+	echo "Adding post-upgrade run-once script to RC" | tee -a /tmp/upgradeLog
+	rc-update add os-upgrade default 2>&1 | tee -a /tmp/upgradeLog
+	
+	# Save final config before upgrade starts
+	lbu commit 2>&1 | tee -a /tmp/upgradeLog
+	
+	cd /media/$LBU_MEDIA
+	mount -o remount,rw /media/$LBU_MEDIA 2>&1 | tee -a /tmp/upgradeLog
+	cat /tmp/upgradeLog >> upgradeLog
+	rm /tmp/upgradeLog 2>&1 | tee -a upgradeLog
+	wget https://dl-cdn.alpinelinux.org/alpine/latest-stable/releases/${ARCH}/alpine-rpi-${LATEST_RELEASE}-${ARCH}.tar.gz 2>&1 | tee -a upgradeLog
+	wget https://dl-cdn.alpinelinux.org/alpine/latest-stable/releases/${ARCH}/alpine-rpi-${LATEST_RELEASE}-${ARCH}.tar.gz.sha256 2>&1 | tee -a upgradeLog
+	if [ $(sha256sum -c alpine-rpi-${LATEST_RELEASE}-${ARCH}.tar.gz.sha256 | grep "alpine-rpi-${LATEST_RELEASE}-${ARCH}.tar.gz: OK" | wc -l) -eq 1 ]; then
+		echo "Alpine Linux Release checksum confirmed. Proceeding with upgrade..." | tee -a upgradeLog
+	 	# Remove old files which will be replaced when new version is extracted
+	 	rm -r /media/$LBU_MEDIA/apks 2>&1 | tee -a upgradeLog
+	  	rm -r /media/$LBU_MEDIA/boot 2>&1 | tee -a upgradeLog
+	   	rm -r /media/$LBU_MEDIA/overlays 2>&1 | tee -a upgradeLog
+	    	rm /media/$LBU_MEDIA/*.dtb 2>&1 | tee -a upgradeLog
+	     	rm /media/$LBU_MEDIA/*.elf 2>&1 | tee -a upgradeLog
+	      	rm /media/$LBU_MEDIA/*.dat 2>&1 | tee -a upgradeLog
+	       	rm /media/$LBU_MEDIA/bootcode.bin 2>&1 | tee -a upgradeLog
+		rm /media/$LBU_MEDIA/cmdline.txt 2>&1 | tee -a upgradeLog
+	 	rm /media/$LBU_MEDIA/config.txt 2>&1 | tee -a upgradeLog
+		#rm /media/$LBU_MEDIA/apks/$ARCH/* # Clear old apk packages from previous version
+		#rm /media/$LBU_MEDIA/cache/*
+		#apk update | tee -a upgradeLog
+		#apk update | tee -a upgradeLog # Needs to be run twice?
+		#apk cache -v download | tee -a upgradeLog
+		tar xzf alpine-rpi-${LATEST_RELEASE}-${ARCH}.tar.gz 2>&1 | tee -a upgradeLog
+		rm alpine-rpi-${LATEST_RELEASE}-${ARCH}.tar.gz alpine-rpi-${LATEST_RELEASE}-${ARCH}.tar.gz.sha256 2>&1 | tee -a upgradeLog
+		echo "Upgrade Complete.  Syncing drive and rebooting..." | tee -a upgradeLog
+		echo "" | tee -a upgradeLog
+		sync 2>&1 | tee -a upgradeLog
+		reboot
+	else
+		echo "DOWNLOADED FILE CHECKSUM FAILURE. ABORTING AND CLEANING UP." | tee -a upgradeLog
+		rm alpine-rpi-$LATEST_RELEASE-$ARCH.tar.gz alpine-rpi-$LATEST_RELEASE-$ARCH.tar.gz.sha256 2>&1 | tee -a upgradeLog
+		
+		rc-update del /etc/init.d/os-upgrade 2>&1 | tee -a upgradeLog
+		lbu exclude /etc/init.d/os-upgrade /bin/os-upgrade.sh 2>&1 | tee -a upgradeLog
+		rm /etc/init.d/os-upgrade /bin/os-upgrade.sh 2>&1 | tee -a upgradeLog
+	
+		lbu commit 2>&1 | tee -a upgradeLog
+		echo "CLEANUP COMPLETE. EXITING WITH ERROR CODE 1." | tee -a upgradeLog
+		echo "View upgrade log: /media/$LBU_MEDIA/upgradeLog" | tee -a upgradeLog
+		echo "" | tee -a upgradeLog
+		mount -o remount,ro /media/$LBU_MEDIA
+		exit 1
+	fi
+else
+	# SKIP_INSTALL SET - Only Run Finishing Script
+	os-upgrade.sh
 fi
 
 exit 0
